requirements.txt
Django==2.2.10
django-background-tasks==1.1.13
djangorestframework==3.12.1
Markdown==3.2.2
django-filter==2.4.0
google-api-core==1.22.2
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
